Lots of changes for cv++ compability, might break linux build...
git-svn-id: http://svn.springlobby.info/trunk@157 1d9fd447-6eb1-4854-8194-394dbce8a230

+++ src/battlelist.h
@@ -42,7 +42,7 @@ class BattleList
       return m_battles.find(id) != m_battles.end();
     }
 
-    int GetNumBattles() { return m_battles.size(); }
+    int GetNumBattles() { return (int)m_battles.size(); }
  
   private:
     battle_map_t m_battles;
